backend/requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
asyncpg==0.29.0
sqlalchemy==2.0.27
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.7
email-validator==2.1.0
alembic==1.13.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0